Supervised learming - Multiple linear regression
Multiple linear regression is a statistical technique used to model the relationship between a continuous dependent variable and two or more independent variables. It aims to find the best-fitting linear equation that predicts the dependent variable based on the independent variables. This method helps quantify the impact of each independent variable while controlling for the effects of others. Multiple linear regression is widely used in fields such as economics, social sciences, and natural sciences to analyze and predict outcomes when multiple factors are at play.