Supervised learning - Decision Tree
A popular machine learning algorithm, Decision Tree. Use this classification algorithm to build a model from historical data of patients, and their respond to different medications. Then you use the trained decision tree to predict the class of a unknown patient, or to find a proper drug for a new patient.