Naive_Bayes_Demo.ipynb
This Jupyter Notebook file demonstrates the application of the Naive Bayes classifier on two different datasets.
The first part of the code analyzes a "Car.csv" dataset. It loads the data, selects features ('Age' and 'Income') and the target variable ('Car'). The dataset is split into training and testing sets, and a Gaussian Naive Bayes classifier is trained and evaluated for predicting car ownership.

The second part uses the famous Iris dataset from scikit-learn. It loads the Iris dataset, prepares the data by separating features and target ('target'), and then splits it into training and testing sets. Another Gaussian Naive Bayes classifier is trained and evaluated on this dataset for classifying Iris flower species.
Both parts of the code demonstrate how to use Naive Bayes classification for different datasets and calculate the mean absolute error as a performance metric for the classifier.